# api-src/new_swagger.yaml
swagger: '2.0'
info:
  contact:
    name: Cyrus Karsan
  description: Delish API ratings information
  title: Delish API
  version: 1.0.1
host: delish-food-292917.appspot.com
paths:
  /add-doc:
    post:
      consumes:
        - application/json
      parameters:
        - in: body
          name: body
          required: true
          schema:
            properties:
              placeid:
                type: string
            type: object
      responses:
        '200':
          description: Rating added successfully
      description: >-
        Create a new document in the ratings collection of the resturant and
        it's rating
      operationId: add-doc
      summary: Add a new rating to the ratings collection in MongoDB
  /downvote:
    put:
      consumes:
        - application/json
      parameters:
        - in: body
          name: body
          required: true
          schema:
            properties:
              placeid:
                type: string
            type: object
      responses:
        '200':
          description: Rating decremented by 1
      description: 'Given a placeid, decrement its rating by 1'
      operationId: downvote
      summary: Decrement resturant rating by 1
  /get-docs:
    get:
      parameters: []
      responses:
        '200':
          description: A successful response
      operationId: get_docs
      summary: Retrieves all documents in the ratings collection in MongoDB
  /upvote:
    put:
      consumes:
        - application/json
      parameters:
        - in: body
          name: body
          required: true
          schema:
            properties:
              placeid:
                type: string
            type: object
      responses:
        '200':
          description: Rating incremented by 1
      description: 'Given a placeid, update its rating by 1'
      operationId: upvote
      summary: Increment resturant rating by 1
  '/{placeid}':
    delete:
      parameters:
        - description: unique google placeid
          in: path
          name: placeid
          required: true
          type: string
      responses:
        '200':
          description: A successful response
      description: >-
        Given a unique google placeid, delete the document associated with the
        id
      operationId: delete_doc
      summary: Removes document from ratings collection
    get:
      parameters:
        - description: google placeid
          in: path
          name: placeid
          required: true
          type: string
      responses:
        '200':
          description: A successful response
      description: >-
        Given a unique google placeid, return the data associated with the
        document.
      operationId: retrieve_doc
      summary: Retrieves document of given id
tags: []
x-components: {}